Increase query limit when the requested row is higher
Before this, when some row much bigger than pageSize was requested and
no data was loaded yet, multiple queries with pageSize results  would be
done until the requested row is loaded. Now, when the index of the
requested row is larger than the pageSize, the size of the query will be
defined by the index instead.